+++ app/gui/src/dashboard/pages/dashboard/Drive/DriveBar/DriveBarToolbar.tsx
@@ -22,7 +22,7 @@ import {
 import { useUploadFiles } from '#/hooks/backendUploadFilesHooks'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
 import { useOffline } from '#/hooks/offlineHooks'
-import AssetSearchBar from '#/layouts/AssetSearchBar'
+import { AssetSearchBar } from '#/layouts/AssetSearchBar'
 import type { TrashCategory } from '#/layouts/CategorySwitcher/Category'
 import { canTransferBetweenCategories } from '#/layouts/CategorySwitcher/Category'
 import { useCategoriesAPI } from '#/layouts/Drive/Categories'
@@ -327,11 +327,14 @@ function TrashFolderToolbar(props: TrashFolderToolbarProps) {
     category,
     parentId: category.homeDirectoryId,
     refetchInterval: null,
+    labels: null,
+    sortDirection: null,
+    sortExpression: null,
   })
 
   const { data: isEmpty } = useSuspenseQuery({
     ...rootDirectoryQueryOptions,
-    select: (data) => data.length === 0,
+    select: (data) => data.assets.length === 0,
   })
 
   const queryClient = useQueryClient()